+++ bindings/nostr-js/src/event/tag.rs
@@ -207,7 +207,7 @@ impl JsTag {
     #[inline]
     #[wasm_bindgen(js_name = asVec)]
     pub fn as_vec(&self) -> Vec<String> {
-        self.inner.as_vec().to_vec()
+        self.inner.as_slice().to_vec()
     }
 
     /// Consume tag and return array of strings